files to run caflood
Make sure CAFLOOD is downloaded and in the right folder (more info: https://www.exeter.ac.uk/research/centres/cws/resources/caddies/). 

Moreover, this is the data for a folder "EAT8a1" for one wet season in the ABM. One such folder is necessary for every wet season run. For the current ABM and paths, folders need to be numbered according to the cycle of the wet season. Thus, the endings of the folders need to be 1, 3, 5, ... depending on how many years you run. Lastly, the rainfall event modelled each years needs to be adapted.